Moved WebviewError class
